refactor(core): rename timestamp-utils.ts to crd-utils.ts
Disambiguates from time-utils.ts, which holds generic time helpers.
This file's sole export, getCrdUpdatedSeconds, is CRD-specific
(reads the SpecUpdateTimestamp label), so crd-utils.ts better reflects
its scope.
